+++ src/main/java/org/esprit/finovate/services/IBillService.java
@@ -5,7 +5,7 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface IBillService {
-    void payBill(int userId, String reference, double amount) throws SQLException;
+    void payBill(Long userId, String reference, double amount) throws SQLException;
 
-    List<Bill> getBillsByUserId(int userId) throws SQLException;
+    List<Bill> getBillsByUserId(Long userId) throws SQLException;
 }
