+++ src/main/java/org/esprit/finovate/services/ITransactionService.java
@@ -5,14 +5,14 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface ITransactionService {
-    void transferMoney(int senderId, Long numeroCarte, String cin, float amount, String description)
+    void transferMoney(Long senderId, Long numeroCarte, String cin, float amount, String description)
             throws SQLException;
 
-    List<Transaction> getTransactionsByUserId(int userId) throws SQLException;
+    List<Transaction> getTransactionsByUserId(Long userId) throws SQLException;
 
-    float getUserBalance(int userId) throws SQLException;
+    float getUserBalance(Long userId) throws SQLException;
 
-    void logTopUp(int userId, float amount) throws SQLException;
+    void logTopUp(Long userId, float amount) throws SQLException;
 
-    float getDailyTransferTotal(int userId) throws SQLException;
+    float getDailyTransferTotal(Long userId) throws SQLException;
 }
